Make SD::size() return a 64b number
Support reporting proper size for >4GB SD cards.

+++ libraries/SD/src/SD.h
@@ -115,8 +115,8 @@ public:
         return blocksPerCluster() * blockSize();
     }
 
-    size_t size() {
-        return (clusterSize() * totalClusters());
+    uint64_t size() {
+        return ((uint64_t)clusterSize() * (uint64_t)totalClusters());
     }
 
 private:
